+++ src/event/log.rs
@@ -1,0 +1,90 @@
+#![allow(clippy::module_name_repetitions)]
+use super::{common::Match, EventCategory, TriagePolicy, TriageScore, MEDIUM};
+use chrono::{serde::ts_nanoseconds, DateTime, Local, Utc};
+use serde::{Deserialize, Serialize};
+use std::{
+    fmt,
+    net::{IpAddr, Ipv4Addr},
+    num::NonZeroU8,
+};
+
+#[derive(Serialize, Deserialize)]
+pub struct ExtraThreat {
+    #[serde(with = "ts_nanoseconds")]
+    pub time: DateTime<Utc>,
+    pub source: String,
+    pub service: String,
+    pub content: String,
+    pub db_name: String,
+    pub rule_id: u32,
+    pub matched_to: String,
+    pub cluster_id: usize,
+    pub attack_kind: String,
+    pub confidence: f32,
+    pub triage_scores: Option<Vec<TriageScore>>,
+}
+
+impl fmt::Display for ExtraThreat {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "-,-,-,-,-,-,ExtraThreat,{},{},{},{},{},{},{},{},{}",
+            DateTime::<Local>::from(self.time).format("%Y-%m-%d %H:%M:%S"),
+            self.service,
+            self.content,
+            self.db_name,
+            self.rule_id,
+            self.matched_to,
+            self.cluster_id,
+            self.attack_kind,
+            self.confidence,
+        )
+    }
+}
+
+impl Match for ExtraThreat {
+    fn src_addr(&self) -> IpAddr {
+        IpAddr::V4(Ipv4Addr::UNSPECIFIED)
+    }
+
+    fn src_port(&self) -> u16 {
+        0
+    }
+
+    fn dst_addr(&self) -> IpAddr {
+        IpAddr::V4(Ipv4Addr::UNSPECIFIED)
+    }
+
+    fn dst_port(&self) -> u16 {
+        0
+    }
+
+    fn proto(&self) -> u8 {
+        0
+    }
+
+    fn category(&self) -> EventCategory {
+        EventCategory::Reconnaissance
+    }
+
+    fn level(&self) -> NonZeroU8 {
+        MEDIUM
+    }
+
+    fn kind(&self) -> &str {
+        "extra threat"
+    }
+
+    fn source(&self) -> &str {
+        self.source.as_str()
+    }
+
+    fn confidence(&self) -> Option<f32> {
+        Some(self.confidence)
+    }
+
+    fn score_by_packet_attr(&self, _triage: &TriagePolicy) -> f64 {
+        // TODO: implement
+        0.0
+    }
+}